// webapp/src/main/javascript/blomst/src/app/rest.classes.d.ts
export interface UserForm extends AbstractRestEntity {
    "@type": "User";
    nickName: string;
    firstName: string;
    lastName: string;
    birthDate: Date;
    password: string;
    picture: MediaType;
    background: MediaType;
    emails: EmailForm[];
    enabled: boolean;
    language: LanguageForm;
    roles: RoleForm[];
}

export interface ResponseOk extends AbstractSchemaEntity {
    ok: boolean;
}

export interface AbstractRestEntity extends RestEntity, LinksContainer {
    created: Date;
    modified: Date;
    links: { [index: string]: Link };
}
// ...
export interface Locale extends Cloneable, Serializable {
}

export interface NotificationActiveResponse {
    size: number;
}

export interface SystemForm extends AbstractSchemaEntity {
    properties: { [index: string]: any };
}

export interface MovieSearchResponse {
    page: number;
    results: MovieSearchResult[];
    total_results: number;
    total_pages: number;
}

export interface Form<E> extends AbstractSchemaEntity {
    form: E;
}

export interface UserSelect extends IdentitySelect {
    "@type": "User";
    nickName: string;
    firstName: string;
    lastName: string;
    birthDate: Date;
    language: LanguageForm;
}

export interface Link {
    url: string;
    method: string;
    rel: string;
    description: string;
    type: LinkType;
}

export interface MediaType extends AbstractRestEntity {
    name: string;
    lastModified: Date;
    data: string;
    thumbnail: Thumbnail;
}

export interface EmailForm {
    value: string;
    confirmed: boolean;
}

export interface LanguageForm extends AbstractRestEntity {
    "@type": "Language";
    locale: Locale;
    name: string;
}

export interface RoleForm extends AbstractRestEntity {
    "@type": "Role";
    name: string;
    description: string;
}

export interface JsonObject extends JsonNode {
    type: "object";
    properties: { [index: string]: JsonNodeUnion };
    additionalProperties: boolean;
    unevaluatedProperties: boolean;
    required: string[];
    minProperties: number;
    maxProperties: number;
}

export interface AbstractSchemaEntity {
    $schema: JsonObject;
}

export interface RestEntity {
    id: string;
}

export interface LinksContainer {
}
// ...
export interface Thumbnail extends AbstractRestEntity {
    name: string;
    lastModified: Date;
    data: string;
}
// ...
export interface JsonNode {
    type: "object" | "string" | "array" | "boolean" | "image" | "enum" | "integer" | "double" | "float";
    title: string;
    description: string;
    widget: Widget;
    naming: boolean;
    id: boolean;
    readOnly: boolean;
    dirty: boolean;
    visible: boolean;
    visibility: boolean;
    links: { [index: string]: Link };
    validators: { [index: string]: ValidatorUnion };
}
// ...
export interface JsonString extends JsonNode {
    type: "string";
    minLength: number;
    maxLength: number;
    pattern: string;
    format: Format;
    contentMediaType: string;
    contentEncoding: string;
}

export interface JsonInteger extends JsonNumeric {
    type: "integer";
}

export interface JsonDouble extends JsonNumeric {
    type: "double";
}

export interface JsonFloat extends JsonNumeric {
    type: "float";
}

export interface JsonArray extends JsonNode {
    type: "array";
    items: JsonNodeUnion;
    minContains: number;
    maxContains: number;
    uniqueItems: boolean;
}

export interface JsonBoolean extends JsonNode {
    type: "boolean";
}

export interface JsonNull {
}

export interface JsonImage extends JsonNode {
    type: "image";
}

export interface JsonEnum extends JsonNode {
    type: "enum";
    enum: string[];
}

export interface JsonNumeric extends JsonNode {
    type: "integer" | "double" | "float";
    multiplyOf: number;
    maximum: number;
    minimum: number;
}

export type LinkType = "table" | "form";

export type Widget = "no-widget" | "checkbox" | "color" | "date" | "datetime-local" | "email" | "file" | "hidden" | "image" | "month" | "number" | "password" | "radio" | "range" | "reset" | "search" | "submit" | "tel" | "text" | "time" | "url" | "week" | "select" | "form" | "table" | "repeat" | "lazy-multi-select" | "lazy-select" | "lazy-select-name" | "textarea" | "editor" | "like" | "json";

export type Status = "PENDING" | "OK" | "DENIED";

export type Format = "noop" | "date-time" | "time" | "date" | "duration" | "email" | "idn-email" | "hostname" | "idn-hostname" | "ip4" | "ip6" | "uuid" | "uri" | "uri-reference" | "iri" | "iri-reference" | "uri-template" | "json-pointer" | "relative-json-pointer" | "regex" | "big-integer" | "big-decimal";

export type AbstractPostFormUnion = ImagePostForm | VideoPostForm | TextPostForm;

export type ValidatorUnion = DecimalMaxValidator | DecimalMinValidator | DigitsValidator | EmailValidator | FutureOrPresentValidator | FutureValidator | MaxValidator | MinValidator | NegativeOrZeroValidator | NegativeValidator | NotBlankValidator | NotNullValidator | PastOrPresentValidator | PastValidator | PatternValidator | PositiveOrZeroValidator | SizeValidator;

export type JsonNodeUnion = JsonString | JsonInteger | JsonDouble | JsonFloat | JsonObject | JsonArray | JsonBoolean | JsonNull | JsonImage | JsonEnum;